ribbon/requirements.txt
certifi==2019.9.11
coverage==4.5.4
Django==2.2.5
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2
